refactoring : property configuration separation
