+++ src/main/java/org/devcourse/voucher/VoucherApplication.java
@@ -6,8 +6,6 @@ import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-//@ComponentScan(basePackages = {"org.devcourse.voucher.voucher", "org.devcourse.voucher.customer",
-//        "org.devcourse.voucher.error", "org.devcourse.voucher.view", "org.devcourse.voucher.configuration", "org.devcourse.voucher"})
 public class VoucherApplication {
 
     public static void main(String[] args) {
