+++ src/main/java/org/devcourse/voucher/configuration/FilePathProperties.java
@@ -2,8 +2,10 @@ package org.devcourse.voucher.configuration;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
 
 @Configuration
+@PropertySource(value = "classpath:path.yaml", factory = YamlPropertiesFactory.class)
 @ConfigurationProperties(prefix = "path")
 public class FilePathProperties {
     private String blacklist;
