+++ src/main/resources/error.yaml
@@ -1,0 +1,4 @@
+error:
+  invalidCommand: "Invalid command entered."
+  inputNegativeNumbers: "A value of 0 or less has been entered."
+  inputNotNumbers: "You must enter only numbers."
